requirements.txt
google-api-python-client>=2.100.0
feedparser>=6.0.10
requests>=2.31.0
python-dotenv>=1.0.0
googlenewsdecoder>=0.1.2
openai>=1.0.0
